refactor: finish rename
?? I'm assuming this was a vestige of a prior refactor. Finishes the
renaming of the package to match folder structure and semantics.
