+++ config/serialize/serialize_test.go
@@ -1,4 +1,4 @@
-package fsrepo
+package serialize
 
 import (
 	"os"
